Sound implemented
You need to change the project properties, use the fmod files in the
FMOD folder. You only need to change additional dependancies and library
dependancies. Also in SoundSystem.cpp you need to change the sounds
filepath to your own, all the sound files are in the Sounds folder in
ERIN